Remove ambiguous `getindex` method and test for ambiguities
This has the potential to cause breakage, since previous packages
that relied on this will now throw a MethodError. However, there
doesn't seem to be any way to do this without causing a ton of
ambiguities; this can have a cascading effect on other packages that
want to check for ambiguities to avoid inadvertently introducing them.

A workaround is described in the README.