Fix tree printing with non-iterable children
The original design for this code was to not require any indexing
of the object returned by `children` at all, unless the tree was
defined to be indexable. This seems to have gotten lost over time
(or maybe it never worked, since it wasn't tested). Nevertheless
the fix is easy and let's add a test this time :).