WIP/RFC: Major refactor
I've recently needed to work with trees again, which led me to revisit
this package. Unfornately I found some of the API choices I'd originally
made for this package somewhat awkward and hard to extend for what I
needed, so I decided to sit down for a day and try to clean things up.
This is somewhat breaking, but hopefully mostly for things that nobody
is actually using (indexed trees). The core of my unhappiness were the
implicit stacks, my primary concerns being:

- They mix up iteration and indexing state in an unhealthy way
- They are not really first class, so manipulating them, for e.g.
  a search procedure over a tree whose leaves are sorted is somewhat
  hard.

To address this, the implicit stacks are replaced by a new `cursor`
concept with much the same functionality, except that they are more
first class. They are now the exclusive state of TreeIterators and
may also be produced explicitly by other means and then fed back
into the iterators. Semantically, these are simply treeviews which
complete the tree to include both siblinga and parent links. If
the tree already has those, the cursor is simply the tree itself.

I've also taken this opportunity to significantly clean up the `indexed`
tree kind. I think this was a bit of a failed experiment. After
this refactor, there is an `Indexed` view, which behaves much the
same, but has significantly smaller footprint. Before all the
generic functions were duplicated for indices. Now they are simply
the regular tree functions on Indexed{...}. However, I also think
that nobody is actually using these, so I'm somewhat tempted to
remove the concept of indexed trees entirely. Anybody who has
a similar kind of tree can easily do the same thing I'm doing
for `Indexed` here to get the same behavior.