Unbreak Julia 1.0
Version 0.4 got released claiming to support Julia 1.0
(the [compat] is `julia = "1"`) but 1.0 was dropped from CI.
Unfortunately, not only doesn't the package run on
1.0, it doesn't even parse. This means that every package
that wants to update their AbstractTrees dependency might try this

    [compat]
    AbstractTrees = "0.3, 0.4"

and while this works fine on Julia 1.6, it breaks on Julia 1.0.
Had AbstractTrees 0.4 merely declared incompatibility
with Julia 1.0, the package manager would have
handled everything for us---it would install AbstractTrees 0.3.x
on Julia 1.0, and 0.4.x on 1.6 and higher.

Now we are a bit stuck. One option is to see if General
would accept a PR retrospectively making AbstractTrees
0.4 uninstallable on 1.0. The alternative is to provide
at least minimal support for 1.0. I've taken that route here,
but I'd be equally fine with the idea of making a PR to
General.